% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_ethogram}
\alias{compute_ethogram}
\title{Compute Ethogram (Dispatcher)}
\usage{
compute_ethogram(
  data,
  mode = c("auto", "intervals", "samples", "implied"),
  x,
  xend,
  y,
  behaviour,
  group = NULL,
  colour = NULL,
  interval = NULL,
  interval_mode = c("guess", "explicit"),
  remove_nas = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{mode}{One of `"auto"`, `"intervals"`, `"samples"`, `"implied"`.}

\item{x}{Column name for start/sample times.}

\item{xend}{Column name for end times.}

\item{y}{Column name for the y axis.}

\item{behaviour}{Column name for behaviour labels.}

\item{group}{Column name for grouping (optional).}

\item{colour}{Column name for colour (optional).}

\item{interval}{Fixed interval between samples (optional).}

\item{interval_mode}{Use `"explicit"` to require interval or `"guess"` to infer.}

\item{remove_nas}{Remove rows with `NA` behaviour values.}
}
\description{
`r lifecycle::badge("experimental")`
}
