% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggethos.R
\name{compute_ethogram_stat}
\alias{compute_ethogram_stat}
\title{Compute ethogram}
\usage{
compute_ethogram_stat(
  data,
  scales,
  align_trials,
  remove_nas,
  mode = c("auto", "intervals", "samples", "implied"),
  interval = NULL,
  interval_mode = c("guess", "explicit"),
  align_by = NULL,
  align_mode = c("zero")
)
}
\description{
`r lifecycle::badge("experimental")`
}
\keyword{internal}
