% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_implied}
\alias{compute_implied}
\title{Compute Ethogram from Implied Order}
\usage{
compute_implied(
  data,
  y,
  behaviour,
  group = NULL,
  colour = NULL,
  remove_nas = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{y}{Column name for the y axis.}

\item{behaviour}{Column name for behaviour labels.}

\item{group}{Column name for grouping (optional).}

\item{colour}{Column name for colour (optional).}

\item{remove_nas}{Remove rows with `NA` behaviour values.}
}
\description{
`r lifecycle::badge("experimental")`
}
