% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethogram_demo.R
\docType{data}
\name{ethogram_demo}
\alias{ethogram_demo}
\alias{ethogram_demo_duration}
\title{Minimal ethogram demo datasets}
\format{
Each is a tibble
}
\usage{
ethogram_demo

ethogram_demo_duration
}
\description{
Deterministic datasets designed to make ethogram outputs easy to verify.
}
\examples{
ethogram_demo
ethogram_demo_duration
}
\keyword{datasets}
