% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{guess_interval}
\alias{guess_interval}
\title{Guess Axis Interval}
\usage{
guess_interval(diffs)
}
\description{
`r lifecycle::badge("experimental")` This function guesses the interval using differences in the x axis provided to geom_ethogram().
}
\keyword{internal}
