% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wombats.R
\docType{data}
\name{wombats}
\alias{wombats}
\alias{wombats_duration}
\title{Wombats datasets}
\format{Each is a tibble}
\usage{
wombats

wombats_duration
}
\description{
These datasets contain toy data to experiment with
the way `ggethos` expects data to be given
}
\examples{
wombats
wombats_duration
}
\keyword{datasets}
