% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggethos.R
\name{compute_ethogram}
\alias{compute_ethogram}
\title{Compute ethogram}
\usage{
compute_ethogram(data, scales, align_trials, remove_nas)
}
\description{
`r lifecycle::badge("experimental")`
}
\keyword{internal}
