% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggethos.R
\name{geom_ethogram}
\alias{geom_ethogram}
\title{Ethogram Geom Wrapper}
\usage{
geom_ethogram(mapping = NULL, data = NULL, stat = "etho",
  position = "identity", ..., size = 5, arrow = NULL,
  lineend = "butt", linejoin = "round", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, align_trials = FALSE,
  remove_nas = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetics created by `aes()` or
`aes_()`. If specified and `inherit.aes = TRUE` (the
default), it is combined with the default mapping at the top level of the
plot. You must supply `mapping` if there is no plot mapping.}

\item{data}{Data provided for the plot if not provided through previous `ggplot(data, ...)` layer}

\item{align_trials}{boolean indicating whether to align all trials to zero (default = FALSE)}

\item{remove_nas}{boolean indicating whether to remove the `NAs` in the data or not (default = TRUE)}
}
\description{
Ethogram Geom Wrapper
}
\seealso{
[ggplot2::geom_segment()]
}
